+++ package.json
@@ -12,6 +12,7 @@
     "axios": "^1.2.2",
     "bootstrap": "^5.2.3",
     "formik": "^2.2.9",
+    "json-server": "^0.17.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
